// src/types/driverTypes.ts
export interface DriverData {
    name: string;
    phone: string;
    carNumber: string;
    carModel: string;
    carYear: number; // Ensure this is a number
    carType: string;
    drivingLicense: string;
    aadhaarCard: string;
    employmentType: string;
    password: string;
    status: boolean;
    driverType: string;
  }
  